requirements.txt
langchain
langchain-google-genai
python-dotenv
sysv-ipc
vosk
sounddevice
sounddevice
numpy
whisper
pynput
torch
